fix integer division

+++ city-traffic-control/control-system/src/main/java/at/jku/softengws20/group1/controlsystem/service/TrafficStatusRepository.java
@@ -46,7 +46,7 @@ public class TrafficStatusRepository {
         }
     }
 
-    private void setTrafficLoad(RoadSegment roadSegment, int capacity, int carsWaiting) {
+    private void setTrafficLoad(RoadSegment roadSegment, double capacity, double carsWaiting) {
         // traffic load can only be calulated of capacity > 0
         if (capacity != 0 ) {
             double trafficLoad = Math.min(carsWaiting / capacity, 1.0);
